+++ recipes/mongodb_org_repo.rb
@@ -43,7 +43,7 @@ when 'debian'
   apt_repository 'mongodb' do
     uri node['mongodb']['repo']
     distribution "#{node['lsb']['codename']}/mongodb-org/#{package_version_major}"
-    components node['platform'] == 'ubuntu' ? ['multiverse'] : ['main']
+    components platform?('ubuntu') ? ['multiverse'] : ['main']
     key "https://www.mongodb.org/static/pgp/server-#{package_version_major}.asc"
   end
 when 'amazon', 'fedora', 'rhel'